requirements.txt
# Hive Agent Framework - Production Requirements
# ===============================================
# Generated: 2026-01-31
# Python: 3.11+

# Core framework dependencies
pydantic>=2.5.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
pypdf>=4.0.0
pandas>=2.1.4
jsonpath-ng>=1.6.0
fastmcp>=2.0.0
diff-match-patch>=20230430
python-dotenv>=1.0.0

# LLM providers
litellm>=1.50.0

# External integrations
simple-salesforce>=1.12.6
requests>=2.31.0

# Development dependencies (optional)
# pip install hive[dev]
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

# Sandbox dependencies (optional)
# pip install hive[sandbox]
# RestrictedPython>=7.0

# OCR dependencies (optional)
# pip install hive[ocr]
# pytesseract>=0.3.10
# pillow>=10.0.0
